0.58.0
Migrating to D.JS V13, changes had to be made so that the bot can keep working. This update includes updates on codes so that it is compatible for V13. Later on, slash commands will be added for admin commands and possibly all commands due to the convenience of slash commands. This is also the first update that actually is showed to the public. After switching the host, this repository can be public as the token for the bot can be removed.

Acto